+++ client/app/intake/pages/search.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { doFileNumberSearch, setFileNumberSearch } from '../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, RAMP_INTAKE_STATES } from '../constants';
-import { getRampElectionStatus } from '../selectors';
+import { getIntakeStatus } from '../selectors';
 
 const rampIneligibleInstructions = <div>
   <p>
@@ -97,7 +97,7 @@ class Search extends React.PureComponent {
     const {
       searchErrorCode,
       searchErrorData,
-      rampElectionStatus,
+      intakeStatus,
       formType
     } = this.props;
 
@@ -105,7 +105,7 @@ class Search extends React.PureComponent {
       return <Redirect to={PAGE_PATHS.BEGIN} />;
     }
 
-    switch (rampElectionStatus) {
+    switch (intakeStatus) {
     case RAMP_INTAKE_STATES.STARTED:
       return <Redirect to={PAGE_PATHS.REVIEW} />;
     case RAMP_INTAKE_STATES.REVIEWED:
@@ -139,8 +139,7 @@ class Search extends React.PureComponent {
 
 export default connect(
   (state) => ({
-    rampElection: state.rampElection,
-    rampElectionStatus: getRampElectionStatus(state),
+    intakeStatus: getIntakeStatus(state),
     fileNumberSearchInput: state.intake.fileNumberSearch,
     fileNumberSearchRequestStatus: state.intake.requestStatus.fileNumberSearch,
     searchErrorCode: state.intake.searchErrorCode,
